+++ charts/zitadel/acceptance/accessibility.go
@@ -3,6 +3,10 @@ package acceptance
 import (
 	"context"
 	"fmt"
+	mgmt_client "github.com/zitadel/zitadel-go/v2/pkg/client/management"
+	"github.com/zitadel/zitadel-go/v2/pkg/client/middleware"
+	"github.com/zitadel/zitadel-go/v2/pkg/client/zitadel"
+	mgmt_api "github.com/zitadel/zitadel-go/v2/pkg/client/zitadel/management"
 	"net/http"
 	"strings"
 	"sync"
@@ -12,12 +16,22 @@ import (
 	corev1 "k8s.io/api/core/v1"
 )
 
-type checkOptions struct {
+type checkOptions interface {
+	execute(ctx context.Context) error
+}
+
+type checkOptionsFunc func(ctx context.Context) error
+
+func (f checkOptionsFunc) execute(ctx context.Context) error {
+	return f(ctx)
+}
+
+type httpCheckOptions struct {
 	getUrl string
 	test   func(response *http.Response, body []byte) error
 }
 
-func (c *checkOptions) execute(ctx context.Context) (err error) {
+func (c *httpCheckOptions) execute(ctx context.Context) (err error) {
 	checkCtx, checkCancel := context.WithTimeout(ctx, 5*time.Second)
 	defer checkCancel()
 	//nolint:bodyclose
@@ -34,30 +48,51 @@ func (c *checkOptions) execute(ctx context.Context) (err error) {
 func (s *ConfigurationTest) checkAccessibility(pods []corev1.Pod) {
 	ctx, cancel := context.WithTimeout(s.Ctx, time.Minute)
 	defer cancel()
+	aud := Audience()
 	tunnels := []interface{ Close() }{CloseFunc(ServiceTunnel(s))}
 	defer func() {
 		for _, t := range tunnels {
 			t.Close()
 		}
 	}()
-	checks := append(zitadelStatusChecks(8080), &checkOptions{
-		getUrl: "http://localhost:8080/ui/console/assets/environment.json",
-		test: func(resp *http.Response, body []byte) error {
-			if err := checkHttpStatus200(resp, body); err != nil {
-				return err
-			}
-			bodyStr := string(body)
-			for _, expect := range []string{
-				`"api":"http://localhost:8080"`,
-				`"issuer":"http://localhost:8080"`,
-			} {
-				if !strings.Contains(bodyStr, expect) {
-					return fmt.Errorf("couldn't find %s in environment.json content %s", expect, bodyStr)
+	checks := append(
+		zitadelStatusChecks(8080),
+		&httpCheckOptions{
+			getUrl: Audience() + "/ui/console/assets/environment.json",
+			test: func(resp *http.Response, body []byte) error {
+				if err := checkHttpStatus200(resp, body); err != nil {
+					return err
 				}
-			}
-			return nil
+				bodyStr := string(body)
+				for _, expect := range []string{
+					fmt.Sprintf(`"api":"%s"`, aud),
+					fmt.Sprintf(`"issuer":"%s"`, aud),
+				} {
+					if !strings.Contains(bodyStr, expect) {
+						return fmt.Errorf("couldn't find %s in environment.json content %s", expect, bodyStr)
+					}
+				}
+				return nil
+			},
 		},
-	})
+		checkOptionsFunc(func(ctx context.Context) error {
+			conn, err := mgmt_client.NewClient(
+				Audience(),
+				fmt.Sprintf("%s:%d", Domain, Port),
+				[]string{zitadel.ScopeZitadelAPI()},
+				zitadel.WithJWTProfileTokenSource(middleware.JWTProfileFromFileData([]byte(`{"type":"serviceaccount","keyId":"229185755715993707","key":"-----BEGIN RSA PRIVATE KEY-----\nMIIEpAIBAAKCAQEAm8bpVfzWJuZsEz1VfTrwSAdkbH+i/u2NS4dv60lwIjtXzrU7\n1xZkHw9jxqz+c+APTaTzp1KY49Dc/wcwXv032FuD1GK2ZSRnMaHm8QnNt8Xhi0e8\nBlu3QQmlqxWPCI67wDPUwXoSHM+r9gQXn2pOR0oonoLP+Gzef+RRj1zUFpZmHWPX\nxw4UWWHwl4xChw9iyO4HbZZGe6wBVYVWe2BnvviCVEeKapyjaCqokZES38S4+S2X\nit202xLlRDyXs3XFWmBzHGmEsxx3LZZor85Kbph/bGjDcV8rdQC1YKC++z8OhuLp\n79GltP7YWrfMN3Z8iRUJQY9APrKQYtljVkWrnQIDAQABAoIBAQCIRZrLyRHCF+LF\ndes6UPvv1t+n9oQtRLxNLV7f0m+Q0p7+yhZeE01kyn67R4yU65YXk0w+vIfZC1a4\nlp5fCl73Gx+ZBP2QPyczCPHRPIVE1Yt33zoByevmrjzKDGMC1nIyMmVVF6eOorFI\n1s2ffEycGqir+b1bEkoWUTJ0Gn3Cf1PE4vTgenHhCrYSvMsbmszQ5GDlfxNj27qf\nF2YrnLx11GplMYU0YEzGqSQHxw76rrmF7yiTvbB+olsjXWARAJxBriSlrF2BDYQk\n+HJ8MEwhWhncaZH1i0Xz/jarDBizpo2o1+K1ZqF6RBUknT72EPnMxI9JsvS4FH44\nZfbrujBhAoGBAMQnx6tO79GpnBIAr7iELyUu5F4mCdU6D0rOAiCjXPpCUAdCDuwX\nzROonIGXPPmhzXXtxebeTz4cf+P8p6tUnrqpl/f0Oi1DMOzv0jL/SAUDC9uUrg6k\nurXZT2dgeONwd1pADyNXSpbZfwRE5IoecFg6cgFi4kune0mdG3mr8QjpAoGBAMtN\nerrMc+4bc3GsmWG4FSXn3xlWMeVGIo2/owP2P5MuMu0ibjofZkl28y0xo8dJgWmv\nLiFSEOhUy+TXZK7K1a2+fD+AXHHaHkBjNbTmCaAbf7rZnuUL4iZVpQyIoTCVuAwo\nC6bsE4TcwGddk4yZj/WZ7v1be+uNgeYwQr2UshyVAoGAN8pYsBCzhR6IlVY8pG50\nOk8sBNss0MjCsLQHRuEwAL37pRTUybG7UmwSl4k8foPWvEP0lcWFJFVWyrGBvulC\nfDTgVFXSdi02LS3Iy1hwU3yaUsnm96NCt5YnT2/Q8l96kuDFbXfWbzFNPxmZJu+h\nZHa7FknZs0rfdgCJYAHXfIECgYEAw3kSqSrNyMICJOkkbO2W/+RLAUx8GwttS8dX\nkQaip/wCoTi6rQ3lxnslY23YIFRPpvL1srn6YbiudrCXMOz7uNtvEYt01082SQha\n6j1IQfZOwLRfb7EWV29/i2aPPWynEqEqWuuf9N5f7MLvjH9WCHpibJ4aryhXHqGG\nekvPWWUCgYA5qDsPk5ykRWEALbunzB/RkpxR6LTLSwriU/OzRswOiKo8UPqH4JZI\nOsFAgudG5H+UOEGMuaSvIq0PLbGex16PjKqUsRwgIoPdH8183f9fxZSJDmr7ELIy\nZJEvE3eJnYwMOpSEZS0VR5Sw0CmKV2Hhd+u6rRB8YjXMP0nAVg8eOA==\n-----END RSA PRIVATE KEY-----\n","userId":"229185755715600491"}`))),
+				zitadel.WithInsecure(),
+			)
+			if err != nil {
+				return fmt.Errorf("couldn't create gRPC management client: %w", err)
+			}
+			_, err = conn.Healthz(ctx, &mgmt_api.HealthzRequest{})
+			// TODO: Why is the key checked on the healthz RPC?
+			if strings.Contains(err.Error(), "Errors.AuthNKey.NotFound") {
+				err = nil
+			}
+			return err
+		}))
 	for i := range pods {
 		pod := pods[i]
 		port := 8081 + i
@@ -75,17 +110,20 @@ func (s *ConfigurationTest) checkAccessibility(pods []corev1.Pod) {
 	wait(ctx, s.T(), &wg, "accessibility")
 }
 
-func zitadelStatusChecks(port int) []*checkOptions {
-	return []*checkOptions{{
-		getUrl: fmt.Sprintf("http://localhost:%d/debug/validate", port),
-		test:   checkHttpStatus200,
-	}, {
-		getUrl: fmt.Sprintf("http://localhost:%d/debug/healthz", port),
-		test:   checkHttpStatus200,
-	}, {
-		getUrl: fmt.Sprintf("http://localhost:%d/debug/ready", port),
-		test:   checkHttpStatus200,
-	}}
+func zitadelStatusChecks(port int) []checkOptions {
+	return []checkOptions{
+		&httpCheckOptions{
+			getUrl: fmt.Sprintf("%s://%s:%d/debug/validate", Scheme, Domain, port),
+			test:   checkHttpStatus200,
+		},
+		&httpCheckOptions{
+			getUrl: fmt.Sprintf("%s://%s:%d/debug/healthz", Scheme, Domain, port),
+			test:   checkHttpStatus200,
+		},
+		&httpCheckOptions{
+			getUrl: fmt.Sprintf("%s://%s:%d/debug/ready", Scheme, Domain, port),
+			test:   checkHttpStatus200,
+		}}
 }
 
 func checkHttpStatus200(resp *http.Response, _ []byte) error {
