+++ examples/2-postgres-secure/zitadel-values.yaml
@@ -2,6 +2,7 @@ zitadel:
   masterkey: x123456789012345678901234567891y
   configmapConfig:
     ExternalSecure: false
+    ExternalDomain: zitadel.default.127.0.0.1.sslip.io
     TLS:
       Enabled: false
     Database:
