+++ examples/5-referenced-secrets/zitadel-values.yaml
@@ -2,6 +2,7 @@ zitadel:
   masterkeySecretName: existing-zitadel-masterkey
   configmapConfig:
     ExternalSecure: false
+    ExternalDomain: zitadel.default.127.0.0.1.sslip.io
     TLS:
       Enabled: false
     Database:
